Added ability to "draw" and change grid size
Changed cell color by adding a class when event listener for each cell is activated. Moved the code that made the grid into a function. Also added a delete function for every new function call triggered when the button is clicked, so that it clears the current grid when the function is called except for the first call when it creates the original grid users see when the website first loads. Added a button event listener which triggers the function.